fix: restructered code and folder structure
