+++ src/services/users/donations.services.js
@@ -3,14 +3,11 @@ import models from '../../database/models';
 const { Donation } = models;
 
 export const saveDonation = async (body) => {
-  // {userid: u, amount: a, description: d}
   try {
-    const donation = Donation.create(body)
-      .then((d) => ({ status: 201, data: d }))
-      .catch((e) => ({ status: 400, data: e }));
-    return donation;
+    const donation = await Donation.create(body);
+    return { status: 201, data: donation };
   } catch (error) {
-    return error;
+    return { status: 409, data: error };
   }
 };
